+++ NEWS.md
@@ -2,11 +2,15 @@
 
 #### ✨ features and improvements
 
+  * `rearrangeFiles()` no longer overwrites time locale
+
 #### 🐛 bug fixes
 
 #### 💬 documentation etc
 
 #### 🍬 miscellaneous
+
+  * Extend code coverage
 
 
 # gimms 1.2.5 (2026-03-21)